+++ src/services/books.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Book } from '../entities/book.entity';
@@ -16,6 +16,7 @@ export class BooksService {
     const book: Book = new Book();
     book.title = createBookDto.title;
     book.copies_available = createBookDto.copies_available;
+    book.total_copies = createBookDto.total_copies;
     book.author_id = createBookDto.author_id;
     book.category_id = createBookDto.category_id;
     book.isbn = createBookDto.isbn;
@@ -24,25 +25,40 @@ export class BooksService {
   }
 
   findAllBook(): Promise<Book[]> {
-    return this.bookRepository.find();
+    const books = this.bookRepository.find();
+    if (!books) {
+      throw new NotFoundException('Books not found.');
+    }
+    return books;
   }
 
-  findOneBook(id: number) {
+  async findOneBook(id: number) {
+    const book = await this.findOneBook(id);
+    if (!book) {
+      throw new NotFoundException('Book not found.');
+    }
     return this.bookRepository.findOneBy({ id });
   }
 
-  updateBook(id: number, updateBookDto: UpdateBookDto): Promise<Book> {
-    const book: Book = new Book();
+  async updateBook(id: number, updateBookDto: UpdateBookDto): Promise<Book> {
+    const book = await this.findOneBook(id);
+    if (!book) {
+      throw new NotFoundException('Book not found.');
+    }
     book.title = updateBookDto.title;
     book.author_id = updateBookDto.author_id;
     book.copies_available = updateBookDto.copies_available;
+    book.total_copies = updateBookDto.total_copies;
     book.isbn = updateBookDto.isbn;
     book.publication_year = updateBookDto.publication_year;
-    book.id = id;
     return this.bookRepository.save(book);
   }
 
-  removeBook(id: number) {
+  async removeBook(id: number) {
+    const book = await this.findOneBook(id);
+    if (!book) {
+      throw new NotFoundException('Book not found.');
+    }
     return this.bookRepository.delete({ id });
   }
 
@@ -52,15 +68,15 @@ export class BooksService {
     if (book) {
       book.copies_available -= 1;
       this.bookRepository.save(book);
+    } else {
+      throw new NotFoundException('Book not found.');
     }
   }
 
-  async increaseCopies(book_id: number): Promise<void> {
-    const book: Book = await this.findOneBook(book_id);
-
-    if (book) {
-      book.copies_available += 1;
-      this.bookRepository.save(book);
-    }
+  async updateCopiesAvailable(
+    book_id: number,
+    copies_available: number,
+  ): Promise<void> {
+    await this.bookRepository.update(book_id, { copies_available });
   }
 }
